refactor(android/engine): Update namespaces
